More cleanup and restructuring.
- collapse 3 layers of API -> S3 API -> S3 Request to API -> S3 Request.
- Move/Rename API's into their own file.
- Improve overall readability of the code.
