+++ api-s3-definitions.go
