+++ api_handlers_test.go
@@ -21,20 +21,8 @@ import (
 	"io"
 	"net/http"
 	"strconv"
-	"strings"
 	"time"
 )
-
-type userAgentHandler struct{}
-
-func (h userAgentHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	// Verify if new user agents are set for the same client.
-	if strings.Contains(r.Header.Get("User-Agent"), "new-hello-app") {
-		w.WriteHeader(http.StatusOK)
-		return
-	}
-	w.WriteHeader(http.StatusBadRequest)
-}
 
 // bucketHandler is an http.Handler that verifies bucket
 // responses and validates incoming requests.
